requirements.txt
Flask==2.2.2
google-cloud-datastore>=1.15.3,<2.0
google-auth==2.6.2
requests==2.27.1
WTForms==3.0.1
Flask-WTF==1.0.1
python-slugify>=5.0.0
google-cloud-ndb
flask-login==0.6.2
Werkzeug
email-validator==1.3.0
